Split driver library
The driverlibrary now is the repository of drivers for use with the
sensor or userdriver frameworks.

The driversupport now has the boardDefault code and may contain additional
support code over time.
