+++ accel_on_lcd/build.gradle
@@ -44,6 +44,7 @@ android {
 
 dependencies {
     compile 'io.mraa.at.upm:upm_jhd1313m1:1.1.0'
+    compile project(':driversupport')
     compile project(':driverlibrary')
     provided 'com.google.android.things:androidthings:0.3-devpreview'
 }
