+++ driverlibrary/README.md
@@ -6,14 +6,6 @@ The files in this library are gathered here for convenience... they can be copie
 the examples or split into seperate libraries as desired.
 
 
-BoardDefaults.java
-------------------
-This code determines the SOM and carrier board that is being used based on the board information
-provided by the system and the characteristics of the board (Edison Arduino has a GPIO expander).
-
-Support for additonal SOMs and carrier boards can easily be added to this file.
-
-
 Mma7660AccelerometerDriver.java
 -------------------------------
 This code is a class that can be used to register with the Android Things sensor manager. It
