+++ driversupport/README.md
@@ -1,0 +1,6 @@
+BoardDefaults.java
+------------------
+This code determines the SOM and carrier board that is being used based on the board information
+provided by the system and the characteristics of the board (Edison Arduino has a GPIO expander).
+
+Support for additonal SOMs and carrier boards can easily be added to this file.
