+++ driversupport/build.gradle
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion 24
-    buildToolsVersion '25.0.0'
+    buildToolsVersion "25.0.0"
 
     defaultConfig {
-        applicationId "com.example.upm.androidthings.driversamples"
         minSdkVersion 24
         targetSdkVersion 24
         versionCode 1
         versionName "1.0"
-        jackOptions {
-            enabled true
-        }
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
     }
     buildTypes {
         release {
@@ -43,20 +42,7 @@ android {
 }
 
 dependencies {
-    compile 'io.mraa.at.upm:upm_jhd1313m1:1.1.0'
-    compile project(':driversupport')
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'io.mraa.at:mraa:1.6.1'
     provided 'com.google.android.things:androidthings:0.3-devpreview'
-}
-
-repositories {
-    jcenter()
-}
-
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.1'
-    }
 }
