+++ driversupport/src/main/java/com/example/upm/androidthings/driversupport/BoardDefaults.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.example.upm.androidthings.driverlibrary;
+package com.example.upm.androidthings.driversupport;
 
 import android.content.Context;
 import android.content.res.Resources;
 import android.os.Build;
-import mraa.mraa;
+import mraa.*;
 
 public class BoardDefaults {
     // returned by Build.device (does not differentiate the carrier boards)
