+++ driversupport/src/main/res/values/strings.xml
@@ -16,5 +16,7 @@
 -->
 
 <resources>
-    <string name="app_name">driverlibrary</string>
+    <string name="app_name">driversupport</string>
+
+    <string name="GPIO_Edison_Arduino">IO0</string>
 </resources>
