+++ grovebuzzer/src/main/java/com/example/upm/androidthings/driversamples/GroveBuzzer.java
@@ -5,7 +5,7 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 
-import com.example.upm.androidthings.driverlibrary.BoardDefaults;
+import com.example.upm.androidthings.driversupport.BoardDefaults;
 import mraa.mraa;
 import upm_buzzer.Buzzer;
