+++ groveled/src/main/java/com/example/upm/androidthings/driversamples/GroveLed.java
@@ -21,7 +21,7 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 
-import com.example.upm.androidthings.driverlibrary.BoardDefaults;
+import com.example.upm.androidthings.driversupport.BoardDefaults;
 import mraa.mraa;
 
 public class GroveLed extends Activity {
