+++ groverelay/build.gradle
@@ -30,6 +30,6 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':driverlibrary')
+    compile project(':driversupport')
     compile 'io.mraa.at.upm:upm_grove:1.1.0'
 }
