+++ jhd1313m1/src/main/java/com/example/upm/androidthings/driversamples/JhdActivity.java
@@ -22,7 +22,7 @@ import android.os.Handler;
 import android.os.HandlerThread;
 import android.util.Log;
 
-import com.example.upm.androidthings.driverlibrary.BoardDefaults;
+import com.example.upm.androidthings.driversupport.BoardDefaults;
 
 import mraa.mraa;
