+++ mma7660/src/main/java/com/example/upm/androidthings/driversamples/AccelerometerActivity.java
@@ -26,7 +26,7 @@ import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 
 import com.example.upm.androidthings.driverlibrary.Mma7660AccelerometerDriver;
-import com.example.upm.androidthings.driverlibrary.BoardDefaults;
+import com.example.upm.androidthings.driversupport.BoardDefaults;
 import mraa.mraa;
 
 import java.io.IOException;
