+++ tmp006/build.gradle
@@ -30,6 +30,6 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':driverlibrary')
+    compile project(':driversupport')
     compile 'io.mraa.at.upm:upm_tmp006:1.2.0'
 }
